feat(subscriber): add support for `EnvFilter` in `Builder::init` (#337)
`console-subscriber`'s `Builder::init` and `init` functions construct
both a `ConsoleLayer` and a `fmt` layer for logging to stdout.
Currently, the `fmt` layer is always filtered using the `Targets` type
in `tracing-subscriber`. This was chosen because, at the time, the more
sophisticated `EnvFilter` type did not support per-layer filtering.
However, [`tracing-subscriber` v0.3.10][1] added a `Filter`
implementation for `EnvFilter`, allowing it to be used as a per-layer
filter. Therefore, it would be nice to be able to allow using
`EnvFilter` when constructing a subscriber using
`console_subscriber::init`.

This PR adds support for using `EnvFilter` in the `init` functions.
Because the implementation is somewhat more complex than `Targets`,
the use of `EnvFilter` is feature flagged (on by default), so that
`tracing-subscriber`'s `env-filter` can be disabled. When the
`env-filter` feature is disabled, we use `Targets` instead. This way,
users can avoid the regex dependency required by `EnvFilter`.

[1]: https://github.com/tokio-rs/tracing/releases/tag/tracing-subscriber-0.3.10