fix(console): always log to a file instead of `stderr` (#340)
## Motivation

Currently, the `tokio-console` CLI includes `tracing` support for
internal debugging. By default, the logs are written to `stderr`, with
the expectations that users will redirect them to a file if logging is
enabled. 

This isn't great: it means that if a `RUST_LOG` env var is set, but the
user didn't intend to enable the console's logging, logs will be written
to the terminal *while displaying the console TUI*. This results in the
TUI itself getting all messed up (see #339).

If support for `systemd-journald` is enabled, logs are instead written
to the system journal, to avoid having to use this redirection dance. We
should probably change our `tracing` configuration so logs are never
written to stderr/stdout, regardless of whether `journald` is enabled.

## Solution

This branch changes the console so that it logs to a file when
`journald` support is not enabled, rather than to stderr. The log file
location can be overridden from the CLI and config file, and defaults to
`/tmp/tokio-console/logs`. This should fix logging messing up the TUI.

Fixes #339

Signed-off-by: Eliza Weisman <eliza@buoyant.io>