feat(api): add optional histogram outlier details (#351)
Currently, the protobuf `TaskDetails` message sends only the raw
binary representation of an `hdrhistogram` histogram serialized with the
V2 serializer.

When we add an upper bound to our histograms to solve issue #350,
though, we may want to additionally record data about how many outliers
exceeded the upper bound. This branch modifies the `TaskDetails` message
to send a new `DurationHistogram` message, which includes the binary
representation of the histogram in addition to an outlier count and the
value of the highest outlier (if there was one).

This change was performed by changing the histogram field to a `oneof`,
so clients using the new API are still compatible with previous
`console-subscriber` versions that send only the raw histogram.

BREAKING CHANGE:

This is a breaking change *to the Rust bindings* (the `console-api`
crate) due to changing a field from an `Option` to a protobuf `oneof`
(introducing a new enum type). This is **not** a breaking change to the
protobufs themselves --- the actual wire format change is
backwards-compatible, but the generated Rust code changes in a breaking
way.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>