fix(subscriber): fix memory leak from resizing histograms (#351)
## Motivation

Currently, the console uses the [`hdrhistogram` crate][1] to record poll
duration histograms. We currently construct our histograms with
`HdrHistogram::new`, [which _does not_ bound the maximum memory use of
the histogram][2]. This means that a task may use increasingly large
amounts of memory to store its poll duration histogram. When a program
contains tasks that run for the program's entire lifetime, this can
result in those tasks' histograms using increasingly large amounts of
memory.

## Solution

This branch changes `console-subscriber` to bound the maximum value of
the histogram, and report outliers when a duration higher than the bound
is recorded.

Fixes #350

[1]: https://docs.rs/hdrhistogram/latest/hdrhistogram/
[2]: https://github.com/HdrHistogram/HdrHistogram_rust#recording-samples