feat(console): display outliers in histogram view (#351)
This commit updates the console UI to display the outlier count and
highest outlier under the histogram in the task details view, if there
are any outliers. Otherwise, the outlier count is not shown.

For example, here's what the UI looks like if I run the example app and
set the maximum duration to 100 microseconds:

![image](https://user-images.githubusercontent.com/2796466/169893884-5325e1af-7269-4747-93c4-9f1d8903751b.png)