feat(console): dynamically size task details field list
## Motivation

Currently, the task details view always gives the list of fields 50% of
the vertical space. This isn't a particularly effective use of space
when a task has only a small number of fields --- we end up wasting a
lot of the vertical real estate that could be used for something nicer.

## Solution

This branch changes the layout of the task details view to give the list
of fields a `Min` layout constraint with the length of the list of
fields plus the border width. This way, when the list of fields is
short, we don't waste a bunch of vertical space on blank lines.