refac(console): factor out a `Store` type
Currently, there's a bit of repeated boilerplate code for processing
updates to the stored state for tasks, resources, and async ops, such as
tracking a list of new items since the last update, and handling ID
remapping. This PR builds upon the refactor in #358, and replaces the
`state::id` module with a new `state::store` module. This module
contains all of the old code for ID remapping along with a new `Store`
type that implements a store of objects by ID. The `Store` type now
implements much of the boilerplate code that was previously repeated in
the `state::tasks`, `state::resources`, and `state::async_ops` modules.