chore: add a `cargo xtask` for generating protobuf bindings
This is necessary as the current approach, regenerating the bindings in
an integration test, does not work when the protos fail to compile
(which they do after the Tonic update). Since running the crate's tests
requires compiling the crate, if the proto bindings don't compile, we
can't re-run the test.

Therefore, I've added a [`cargo xtask`][xtask] that regenerates the
protobuf bindings.

[xtask]: https://github.com/matklad/cargo-xtask