fix: work around Rust 1.69 almost_swapped lint
The enhanced `clippy::almost_swapped` lint which shipped in Rust 1.69.0
was triggering in some cases where we derive `clap::Subcommand`. It
wasn't possible to allow it, because a deny on `clippy::correctness`
(which implies `clippy::almost_swapped`) was present in the generated
code.

As a result, we had to fix the version of Rust used for Clippy to
1.68.2.

The deny on `clippy::correctness` has been removed from `clap_derive`
3.2.24, allowing us to work around this lint by adding an allow for
`clippy::almost_swapped` at the top of each file where it gets
triggered.

This change updates clap to the latest 3.x version, including
clap_derive and clap_lex. It also adds the necessary allow attributes.