refactoring the help view
Created a trait which generates the help text. This is implemented by
each struct that previously implemented the same `render_help_content`
function. This cleans up the code and allows us to only build the help
view when we actually need it.

Also changed the table view help text to be multi-line. This makes it
easier to read (in my honest opnion).