chore(release): generate changelogs with `git-cliff` (#416)
We maintain separate changelogs for the `tokio-console`, `console-api`,
and `console-subscriber` crates.

Currently, changelogs are generated using the [`clog`] CLI tool.
However, `clog` does not support generating separate changelogs for
different subdirectories of a repo, so the current release process for a
given crate requires generating a changelog for the whole repo since the
last tag, and then manually editing it to remove changes to other
crates. This is a pain.

The [`git-cliff`] changelog-generating tool supports generating a
changelog that only includes commits that modified specific paths in the
repo. This means it can be used to generate separate changelogs for each
crate automatically, unlike `clog`. It's also much more configurable,
and we can use it to generate changelogs that have somewhat nicer
formatting (IMO).

This branch changes changelog generation from using `clog` to use
`git-cliff`, and regenerates all the existing changelogs. The included
`bin/update-changelog.sh` shell script updates the changelog for a given
crate to the provided tag. This will make the release process much less
complicated. 

A subsequent PR will add a shell script that handles updating crate
versions, generating the changelog, and publishing the crate and tag, so
that a release can be published by running a single command.

[`clog`]: https://github.com/clog-tool/clog-cli
[`git-cliff`]: https://git-cliff.org/