chore(nix): unpin nixpkgs version in flake
Currently, the Nix `flake.nix` pins the nixpkgs version to
`release-21.11`. This branch of nixpkgs includes `rustc` v1.56.1, which
is too old to build `tokio-console`. This means that Nix users using the
flake cannot actually build the console.

This branch unpins the nixkpgs version, allowing Nix users to build with
the latest `rustc`.