docs(console): revert "update screenshots in README (#419)"
This reverts commit 4f714845f3cda0978b0b1cc850072878ed4f8599. This
commit breaks the existing docs, because it deletes images that are
linked in the current `docs.rs` documentation for `tokio-console`. Also,
we probably ought to resize the images so the text doesn't get scaled.

In the future, when adding new screenshots, let's make sure not to
delete the existing ones, because their GitHub URLs are referenced by
prior versions of the docs (cc @hds).

We should probably figure out a nicer image hosting solution
eventually...