docs(README): update repository README screenshots and help text (#422)
In #421, the screenshots in the `tokio-console` README were updated to
match the current code. The contents of the README also go into the
docs.rs page.

This change updates the two screenshots in the README in the root of the
repository. They now match the UI shown by the current code at the
latest release of `tokio-console`, 0.1.8. The screenshots are a little
bit wider than the ones in the `tokio-console` README as the example
`app` from the `console-subscriber` crate was used as the instrumented
application and the source file paths are longer.

The screenhots are now using relative links directly to the files in the
repository.

Additionally, the help text in the main README has been updated to
reflect what `tokio-console` now produces. This includes additional
descriptive text and also mentions how to connect to a Unix domain
socket.