use duplex stream to connect client and server
Connecting the gRPC client and server via TCP for the test means that
the port must be randomized so that multiple tests can run at the same
time. This creates the possibility for flaky tests if you get unlucky
with the randomization - or requires the introduction of retry logic.

Instead, we use a duplex stream from Tokio as per the Tonic `mock`
example:
https://github.com/hyperium/tonic/blob/1934825ff52bff26bb88b709aee9ac73d3ea51c0/examples/src/mock/mock.rs