receive updates until test finishes
Previously we received a limited number of updates. Assuming that our
test would complete within that time. However a long test may cause this
to fail, or worse be flaky.

Now we advance to a new test step when the test finishes. After every
update on the client, the current test step is checked, once the
`TestFinished` step is reached, another update is received and then the
client returns.