expose the Aggregator instead of a handle to abort it
As discussed in code review, for such a low level api as
`Server::into_parts`, it makes sense to allow the user to spawn the
aggregator where they like, rather than spawning it internally.

Since I couldn't find a way to return and use a `dyn Future` (boxed or
otherwise), the `Aggregator` has been made public with a (async) single
function `run()` which will start it's run loop.