chore: updates for Rust 1.72.0
Clippy was complaining about 3 new things in 1.72.0. Two new lints and
one updated one.

When `futures-core` was added as an explicit dependency on `console-api`
in #453, it was only added to the `console-api` Cargo.toml file. It
wasn't added to the Cargo.lock file shared by the workspace.

Since Rust 1.72.0, Cargo is adding this dependency to Cargo.lock
locally, so it makes sense to check in that change.