chore(subscriber): trace events drain loop stats
The number of each variant of event encountered while draining the mpsc
channel of events sent from the `ConsoleLayer` to the `Aggregator` is
useful debugging information.

The change adds a `DEBUG` level event which is recorded at the end of
each loop draining the channel. It contains the number of each variant
that has been received.

I've written this same code three separate times looking into different
issues, so I decided it was probably worth the small overhead of
aggregating the counts, even when the tracing event isn't output
anywhere.